requirements.txt
telethon==1.40.0
python-dotenv==1.1.0
numpy==2.3.0
openai==1.84.0
scikit-learn==1.7.0
networkx==3.5